+++ adam_sim/entities/__init__.py
@@ -2,6 +2,8 @@ from .configuration import Configuration
 from .point import Point
 from .vector import Vector
 from .system import System
+from .velocity import Velocity
+from .acceleration import Acceleration
 from ..features.adam.entities import AdamInfo
 from ..features.manipulator.entities import Collision, ManipulatorInfo
 from ..features.base.entities import BaseInfo