+++ adam_sim/features/adam/communication/implementations/real_adam_repository/real_adam_repository.py
@@ -1,10 +1,19 @@
+import time
+from paho.mqtt.client import Client
+
 from ....repository import AdamRepository
-from ....entities import AdamInfo
 
 
 class RealAdamRepository(AdamRepository):
-    def init(self, filename: str) -> None:
-        pass
+    def init(self, host: str, port: int, rate: int) -> None:
+        self.client = Client()
+        self.client.connect(host, port)
+
+        self.count: int = 0
+        self.rate: int = rate
 
     def step(self) -> None:
-        pass
+        self.client.publish("mqtt/adam", self.count)
+        self.count += 1
+
+        time.sleep(1.0 / self.rate)