+++ adam_sim/features/adam/communication/implementations/simulated_adam_repository/simulated_adam_repository.py
@@ -1,7 +1,6 @@
 import mujoco
 
 from ....repository import AdamRepository
-from ....entities import AdamInfo
 
 
 class SimulatedAdamRepository(AdamRepository):
@@ -9,9 +8,6 @@ class SimulatedAdamRepository(AdamRepository):
         # Attributes
         self.model = mujoco.MjModel.from_xml_path(filename)  # type: ignore
         self.data = mujoco.MjData(self.model)  # type: ignore
-
-        # Initial pose
-        mujoco.mj_resetDataKeyframe(self.model, self.data, 0)  # type: ignore
 
     def step(self) -> None:
         mujoco.mj_step(self.model, self.data)  # type: ignore