+++ adam_sim/features/manipulator/repository/manipulator_repository.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from ....entities import Configuration, Vector, System
+from ....entities import Configuration, System, Velocity, Acceleration
 from ..entities import Collision, ManipulatorInfo
 
 
@@ -18,11 +18,19 @@ class ManipulatorRepository(ABC):
         pass
 
     @abstractmethod
-    def set_velocity(self, velocity: Vector) -> None:
+    def set_velocity(self, velocity: Velocity) -> None:
         pass
 
     @abstractmethod
-    def get_velocity(self) -> Vector:
+    def get_velocity(self) -> Velocity:
+        pass
+
+    @abstractmethod
+    def set_acceleration(self, acceleration: Acceleration) -> None:
+        pass
+
+    @abstractmethod
+    def get_acceleration(self) -> Acceleration:
         pass
 
     @abstractmethod